Rename base class again and move to different namespace
